refactor: OutputView 중 일부 BridgeGame에서 처리
